pick up prefab, UI, Player Controller Script
added a player controller script that keeps track of cookies found, the score. and a timmer for the level. added in a UI that shows the score, cookies number, and time. Made the pick up into a prefab and added a rotater scrip so that the pickups look more like an object of interest.